fix android keyboard issue
