+++ src/components/KeyboardAvoidingView/index.js
@@ -1,0 +1,26 @@
+/*
+ * This is a KeyboardAvoidingView only enabled for ios && disabled for all other platforms
+ */
+import React from 'react';
+import {View} from 'react-native';
+import PropTypes from 'prop-types';
+
+const propTypes = {
+    children: PropTypes.node,
+};
+const defaultProps = {
+    children: null,
+};
+
+const KeyboardAvoidingView = props => (
+    // eslint-disable-next-line react/jsx-props-no-spreading
+    <View {...props}>
+        {props.children}
+    </View>
+);
+
+KeyboardAvoidingView.propTypes = propTypes;
+KeyboardAvoidingView.defaultProps = defaultProps;
+KeyboardAvoidingView.displayName = 'KeyboardAvoidingView';
+
+export default KeyboardAvoidingView;
